Make Cmode_t an unsigned long long to have more chanmodes on 32 bit archs. Reported by BlackBishop in https://bugs.unrealircd.org/view.php?id=6301 [skip ci]

+++ include/modules.h
@@ -189,7 +189,7 @@ typedef enum BypassChannelMessageRestrictionType {
 #define EXSJ_MERGE		3 /**< SJOIN: Merging of modes, neither won nor lost */
 
 /** Channel mode bit/value */
-typedef unsigned long Cmode_t;
+typedef unsigned long long Cmode_t;
 
 typedef enum CmodeType {
 	CMODE_NORMAL=0,
